GOOD.  CAN USE upstream with only these modifications.
Untested fix for crash problem, the .S word.

There's a good chance this file will stand as-is,
indefinitely.

Of course it'll be edited *anyway* -- for pleasure.

But it (provisionally) appears to be the
minimal port from the upstream Author's
own work, in the present context.

On branch camelforth_tada_test_b

	modified:   forth_edits.c